+++ Stitch/Graph/StitchAI/GraphPrompting/DataModel/StitchAIRequestBodyFormattable.swift
